Create Code
An autonomous robot that follows either a black line in the white area or a white line in the black area. It is able to detect a particular line and keep following it. Optimize to follow the path even with the circular arc type shape or even complex shape path.
It uses six IR sensors, five at the front ( connected in serial ) and one at back. It uses a unique algorithm which detects the  shortest path and makes the Line Follower Robot to follow that path